+++ library/src/androidMain/kotlin/com/lightningkite/kiteui/views/direct/TextField.android.kt
@@ -45,7 +45,7 @@ public actual open class TextInput actual constructor(context: ElementContext) :
                 (if(theme.font.underline) android.graphics.Paint.UNDERLINE_TEXT_FLAG else 0) or
                 (if(theme.font.strikethrough) Paint.STRIKE_THRU_TEXT_FLAG else 0)
         useAllCaps = theme.font.allCaps
-        native.setTextSize(TypedValue.COMPLEX_UNIT_PX, theme.font.size.value.toFloat())
+        native.setTextSize(TypedValue.COMPLEX_UNIT_PX, theme.font.size.value)
         applyAlign(_align ?: Align.Stretch)
     }
